Enable more analyser rules as errors
Also deduplicates a bit of code inside of SeasonPathParser and adds some more tests
