+++ jellyfin.ruleset
@@ -105,6 +105,28 @@
     <Rule Id="CA1851" Action="Error" />
     <!-- error on CA1854: Prefer a 'TryGetValue' call over a Dictionary indexer access guarded by a 'ContainsKey' check to avoid double lookup -->
     <Rule Id="CA1854" Action="Error" />
+    <!-- error on CA1860: Avoid using 'Enumerable.Any()' extension method -->
+    <Rule Id="CA1860" Action="Error" />
+    <!-- error on CA1862: Use the 'StringComparison' method overloads to perform case-insensitive string comparisons -->
+    <Rule Id="CA1862" Action="Error" />
+    <!-- error on CA1863: Use 'CompositeFormat' -->
+    <Rule Id="CA1863" Action="Error" />
+    <!-- error on CA1864: Prefer the 'IDictionary.TryAdd(TKey, TValue)' method -->
+    <Rule Id="CA1864" Action="Error" />
+    <!-- error on CA1865-CA1867: Use 'string.Method(char)' instead of 'string.Method(string)' for string with single char -->
+    <Rule Id="CA1865" Action="Error" />
+    <Rule Id="CA1866" Action="Error" />
+    <Rule Id="CA1867" Action="Error" />
+    <!-- error on CA1868: Unnecessary call to 'Contains' for sets -->
+    <Rule Id="CA1868" Action="Error" />
+    <!-- error on CA1869: Cache and reuse 'JsonSerializerOptions' instances -->
+    <Rule Id="CA1869" Action="Error" />
+    <!-- error on CA1870: Use a cached 'SearchValues' instance -->
+    <Rule Id="CA1870" Action="Error" />
+    <!-- error on CA1871: Do not pass a nullable struct to 'ArgumentNullException.ThrowIfNull' -->
+    <Rule Id="CA1871" Action="Error" />
+    <!-- error on CA1872: Prefer 'Convert.ToHexString' and 'Convert.ToHexStringLower' over call chains based on 'BitConverter.ToString' -->
+    <Rule Id="CA1872" Action="Error" />
     <!-- error on CA2016: Forward the CancellationToken parameter to methods that take one
         or pass in 'CancellationToken.None' explicitly to indicate intentionally not propagating the token -->
     <Rule Id="CA2016" Action="Error" />
